upgrade from deprecated apis
