+++ openshift-knative-operator/pkg/eventing/extension.go
@@ -34,8 +34,9 @@ func (e *extension) Manifests(ke operatorv1alpha1.KComponent) ([]mf.Manifest, er
 }
 
 func (e *extension) Transformers(ke operatorv1alpha1.KComponent) []mf.Transformer {
-	return append([]mf.Transformer{common.InjectCommonLabelIntoNamespace(), common.VersionedJobNameTransform()},
-		monitoring.GetEventingTransformers(ke)...)
+	tf := []mf.Transformer{common.InjectCommonLabelIntoNamespace(), common.VersionedJobNameTransform()}
+	tf = append(tf, common.DeprecatedAPIsTranformers(e.kubeclient.Discovery())...)
+	return append(tf, monitoring.GetEventingTransformers(ke)...)
 }
 
 func (e *extension) Reconcile(ctx context.Context, comp operatorv1alpha1.KComponent) error {
